chore: add multiple users to retirement queue.

+++ openedx/core/djangoapps/user_api/management/commands/retire_user.py
@@ -14,6 +14,8 @@ from ...models import UserRetirementStatus
 
 
 logger = logging.getLogger(__name__)
+unknown_users = []
+users = []
 
 
 class Command(BaseCommand):
@@ -35,55 +37,89 @@ class Command(BaseCommand):
     def add_arguments(self, parser):
         parser.add_argument(
             '--username',
-            required=True,
             type=str,
             help='Username to be retired'
         )
         parser.add_argument(
             '--user_email',
-            required=True,
             type=str,
             help='User email address.'
+        )
+        parser.add_argument(
+            '--user_file',
+            type=str,
+            help='Comma separated file that have username and user_email of the users that needs to be retired'
         )
 
     def handle(self, *args, **options):
         """
         Execute the command.
         """
-
-        username = options['username']
-        user_email = options['user_email']
         try:
-            user = User.objects.get(username=username, email=user_email)
-        except:
+            userfile = options['user_file']
+        except KeyError:
+            userfile = None
+        try:
+            user_name = options['username']
+            useremail = options['user_email']
+        except KeyError:
+            user_name = None
+            useremail = None
+
+        if userfile:
+            try:
+                userinfo = open(userfile, 'r')
+            except Exception as exc:
+                error_message = f'Error while reading file: {exc}'
+                logger.error(error_message)
+                raise CommandError(error_message)  # lint-amnesty, pylint: disable=raise-missing-from
+
+            for record in userinfo:
+                userdata = record.split(',')
+                username = userdata[0].strip()
+                user_email = userdata[1].strip()
+                try:
+                    users.append(User.objects.get(username=username, email=user_email))
+                except Exception:
+                    unknown_users.append({username: user_email})
+        elif user_name and useremail:
+            try:
+                users.append(User.objects.get(username=username, email=user_email))
+            except Exception:
+                unknown_users.append({username: useremail})
+        else:
+            raise CommandError("Please provide user_file or username and user_email parameter when runing command")
+
+        if len(unknown_users) > 0:
             error_message = (
-                'Could not find a user with specified username and email '
+                'Could not find users with specified username and email '
                 'address. Make sure you have everything correct before '
                 'trying again'
             )
             logger.error(error_message)
-            raise CommandError(error_message)  # lint-amnesty, pylint: disable=raise-missing-from
+            raise CommandError(error_message + f': {unknown_users}')  # lint-amnesty, pylint: disable=raise-missing-from
 
         user_model = get_user_model()
 
         try:
             with transaction.atomic():
-                # Add user to retirement queue.
-                UserRetirementStatus.create_retirement(user)
-                # Unlink LMS social auth accounts
-                UserSocialAuth.objects.filter(user_id=user.id).delete()
-                # Change LMS password & email
-                user.email = get_retired_email_by_email(user.email)
-                user.set_unusable_password()
-                user.save()
+                for user in users:
+                    # Add user to retirement queue.
+                    UserRetirementStatus.create_retirement(user)
+                    # Unlink LMS social auth accounts
+                    UserSocialAuth.objects.filter(user_id=user.id).delete()
+                    # Change LMS password & email
+                    user.email = get_retired_email_by_email(user.email)
+                    user.set_unusable_password()
+                    user.save()
 
-                # TODO: Unlink social accounts & change password on each IDA.
-                # Remove the activation keys sent by email to the user for account activation.
-                Registration.objects.filter(user=user).delete()
+                    # TODO: Unlink social accounts & change password on each IDA.
+                    # Remove the activation keys sent by email to the user for account activation.
+                    Registration.objects.filter(user=user).delete()
 
-                # Delete OAuth tokens associated with the user.
-                retire_dot_oauth2_models(user)
-                AccountRecovery.retire_recovery_email(user.id)
+                    # Delete OAuth tokens associated with the user.
+                    retire_dot_oauth2_models(user)
+                    AccountRecovery.retire_recovery_email(user.id)
         except KeyError:
             error_message = f'Username not specified {user}'
             logger.error(error_message)
@@ -93,7 +129,7 @@ class Command(BaseCommand):
             logger.error(error_message)
             raise CommandError(error_message)  # lint-amnesty, pylint: disable=raise-missing-from
         except Exception as exc:  # pylint: disable=broad-except
-            error_message = f'500 error deactivating account {exc}'
+            error_message = f'500 error deactivating account: {exc}'
             logger.error(error_message)
             raise CommandError(error_message)  # lint-amnesty, pylint: disable=raise-missing-from
